refactor: extract indexer/apply.rs (phase 3, step 11)
Moves the write-path of the indexer into a dedicated module:

- file_contributions / stale_keys_for / build_bare_names (pure fns)
- Indexer::parse_file, apply_file_result, apply_workspace_result
- Indexer::apply_contributions, index_content, prewarm_completion_cache
- Indexer::rebuild_bare_name_cache, rebuild_importable_fqns
- Indexer::index_source_paths (additive sourcePaths scan)
- hash_str moved to apply.rs (only used here)

Adds apply_tests.rs with 23 tests (15 new + 8 migrated from inline
indexer.rs test module). Removes the 8 duplicate inline tests.

indexer.rs: 3876 → 3237 lines
