+++ src/indexer.rs
@@ -49,7 +49,7 @@ pub(crate) use self::infer::{
     find_last_dot_at_depth_zero,
 };
 
-mod cache;
+pub(crate) mod cache;
 pub(crate) use self::cache::workspace_cache_path;
 
 mod discover;
@@ -58,6 +58,10 @@ use self::discover::find_source_files_unconstrained;
 mod scan;
 pub use self::scan::resolve_max_files;
 pub const MAX_FILES_UNLIMITED: usize = usize::MAX;
+
+mod apply;
+pub(crate) use self::apply::{file_contributions, stale_keys_for, build_bare_names};
+use self::apply::hash_str;
 
 // Re-export cache/scan items needed by the inline test module below.
 #[cfg(test)]
@@ -92,88 +96,6 @@ pub(crate) struct StaleKeys {
     /// Both aliases: `pkg.Sym` AND `pkg.FileStem.Sym`.
     pub qualified_keys:   Vec<String>,
     pub package:          Option<String>,
-}
-
-// ─── Pure functions ───────────────────────────────────────────────────────────
-
-/// Pure: compute what a parsed file contributes to each index map.
-/// No side effects. Call `Indexer::apply_contributions` to commit.
-pub(crate) fn file_contributions(result: &FileIndexResult) -> FileContributions {
-    let uri_str = result.uri.to_string();
-    let file_stem: Option<String> = result.uri.to_file_path().ok()
-        .and_then(|p| p.file_stem().map(|s| s.to_string_lossy().into_owned()));
-
-    let mut definitions: HashMap<String, Vec<Location>> = HashMap::new();
-    let mut qualified:   HashMap<String, Location>      = HashMap::new();
-
-    for sym in &result.data.symbols {
-        let loc = Location { uri: result.uri.clone(), range: sym.selection_range };
-        definitions.entry(sym.name.clone()).or_default().push(loc.clone());
-        if let Some(ref pkg) = result.data.package {
-            qualified.insert(format!("{pkg}.{}", sym.name), loc.clone());
-            if let Some(ref stem) = file_stem {
-                if *stem != sym.name {
-                    qualified.insert(format!("{pkg}.{stem}.{}", sym.name), loc);
-                }
-            }
-        }
-    }
-
-    let mut packages: HashMap<String, Vec<String>> = HashMap::new();
-    if let Some(ref pkg) = result.data.package {
-        packages.entry(pkg.clone()).or_default().push(uri_str.clone());
-    }
-
-    let mut subtypes: HashMap<String, Vec<Location>> = HashMap::new();
-    for (super_name, class_loc) in &result.supertypes {
-        subtypes.entry(super_name.clone()).or_default().push(class_loc.clone());
-    }
-
-    FileContributions {
-        definitions,
-        qualified,
-        packages,
-        subtypes,
-        file_data: (uri_str.clone(), Arc::new(result.data.clone())),
-        content_hash: (uri_str, result.content_hash),
-    }
-}
-
-/// Pure: compute which keys to remove from each index map when `uri` is re-indexed.
-/// Requires the *old* FileData to know what the file previously contributed.
-pub(crate) fn stale_keys_for(uri: &Url, old_data: &crate::types::FileData) -> StaleKeys {
-    let file_stem: Option<String> = uri.to_file_path().ok()
-        .and_then(|p| p.file_stem().map(|s| s.to_string_lossy().into_owned()));
-
-    let definition_names: Vec<String> = old_data.symbols.iter()
-        .map(|s| s.name.clone())
-        .collect();
-
-    let mut qualified_keys: Vec<String> = Vec::new();
-    if let Some(ref pkg) = old_data.package {
-        for sym in &old_data.symbols {
-            qualified_keys.push(format!("{pkg}.{}", sym.name));
-            if let Some(ref stem) = file_stem {
-                if *stem != sym.name {
-                    qualified_keys.push(format!("{pkg}.{stem}.{}", sym.name));
-                }
-            }
-        }
-    }
-
-    StaleKeys {
-        definition_names,
-        qualified_keys,
-        package: old_data.package.clone(),
-    }
-}
-
-/// Pure: build sorted, deduplicated list of all symbol names from the definitions map.
-pub(crate) fn build_bare_names(definitions: &DashMap<String, Vec<Location>>) -> Vec<String> {
-    let mut names: Vec<String> = definitions.iter().map(|e| e.key().clone()).collect();
-    names.sort_unstable();
-    names.dedup();
-    names
 }
 
 pub struct Indexer {
@@ -325,376 +247,6 @@ impl Indexer {
         self.live_lines.insert(uri.to_string(), lines);
     }
 
-    // ────────────────────────────────────────────────────────────────────────
-    // Pure Parsing Functions (SOLID refactoring - no side effects)
-    // ────────────────────────────────────────────────────────────────────────
-
-    /// Parse a single file and return structured result data (pure function).
-    /// 
-    /// This is the testable, side-effect-free core of indexing.
-    /// It takes content and returns data without mutating any shared state.
-    /// 
-    /// Use `apply_file_result()` to merge the result into the index.
-    pub fn parse_file(uri: &Url, content: &str) -> FileIndexResult {
-        let data = parser::parse_by_extension(uri.path(), content);
-        let hash = hash_str(content);
-        
-        // Extract supertype relationships for goToImplementation
-        let mut supertypes = Vec::new();
-        let class_kinds = [
-            SymbolKind::CLASS, SymbolKind::INTERFACE, SymbolKind::STRUCT,
-            SymbolKind::ENUM, SymbolKind::OBJECT,
-        ];
-        
-        for sym in &data.symbols {
-            if !class_kinds.contains(&sym.kind) { continue; }
-            let start = sym.selection_range.start.line as usize;
-            let limit = (start + 10).min(data.lines.len());
-            let mut decl_lines: Vec<String> = Vec::new();
-            for line in &data.lines[start..limit] {
-                decl_lines.push(line.clone());
-                if line.contains('{') { break; }
-            }
-            let supers = crate::resolver::extract_supers_from_lines(&decl_lines);
-            let class_loc = Location { uri: uri.clone(), range: sym.selection_range };
-            for super_name in supers {
-                supertypes.push((super_name, class_loc.clone()));
-            }
-        }
-        
-        FileIndexResult {
-            uri: uri.clone(),
-            data,
-            supertypes,
-            content_hash: hash,
-            error: None,
-        }
-    }
-
-    /// Coordinator: apply a single file parse result to the index.
-    ///
-    /// Uses pure `stale_keys_for` to compute removals and `file_contributions`
-    /// to compute insertions. This is the per-file delta path (live edits, on_open).
-    pub fn apply_file_result(&self, result: &FileIndexResult) {
-        let uri_str = result.uri.to_string();
-
-        // ── Remove stale entries ──────────────────────────────────────────────
-        if let Some(old) = self.files.get(&uri_str) {
-            let stale = stale_keys_for(&result.uri, &old);
-            for name in &stale.definition_names {
-                if let Some(mut locs) = self.definitions.get_mut(name) {
-                    locs.retain(|l| l.uri.as_str() != uri_str.as_str());
-                }
-            }
-            for key in &stale.qualified_keys {
-                self.qualified.remove(key);
-            }
-            if let Some(ref pkg) = stale.package {
-                if let Some(mut uris) = self.packages.get_mut(pkg) {
-                    uris.retain(|u| u != &uri_str);
-                }
-            }
-            for mut entry in self.subtypes.iter_mut() {
-                entry.value_mut().retain(|l| l.uri.as_str() != uri_str.as_str());
-            }
-        }
-
-        // ── Insert fresh contributions ────────────────────────────────────────
-        let contrib = file_contributions(result);
-        self.apply_contributions(contrib);
-    }
-
-    /// Coordinator: apply workspace indexing results to the index.
-    ///
-    /// Full-replace path: resets all index maps first, then inserts all file
-    /// contributions. Cache hits are already converted to FileIndexResult by
-    /// `cache_entry_to_file_result` (supertypes included).
-    pub fn apply_workspace_result(&self, result: &WorkspaceIndexResult) {
-        log::info!(
-            "Applying workspace results: {} files parsed, {} cache hits",
-            result.stats.files_parsed, result.stats.cache_hits
-        );
-
-        // Full replace — clear stale state from any previous root or run.
-        self.reset_index_state();
-
-        for file_result in &result.files {
-            let contrib = file_contributions(file_result);
-            self.apply_contributions(contrib);
-        }
-
-        self.rebuild_bare_name_cache();
-
-        log::info!(
-            "Index ready: {} symbols from {} files",
-            self.definitions.len(), self.files.len()
-        );
-    }
-
-    /// Index all configured `sourcePaths` additively — without clearing the workspace index.
-    ///
-    /// Files outside the workspace root are marked as library sources in `library_uris`:
-    /// they contribute to hover, definition, and autocomplete but are excluded from
-    /// findReferences and rename. Files inside the workspace root are indexed but not
-    /// marked as library (they are already covered by the workspace scan; sourcePaths
-    /// can override ignorePatterns for those).
-    ///
-    /// Generation-safe: captures `root_generation` at the start and discards results
-    /// if it changes during async I/O (root switch / explicit reindex).
-    pub async fn index_source_paths(self: Arc<Self>, workspace_root: PathBuf) {
-        let raw_paths = self.source_paths_raw.read().unwrap().clone();
-        if raw_paths.is_empty() { return; }
-
-        let gen = self.root_generation.load(Ordering::SeqCst);
-
-        // Resolve raw paths against workspace root at call time.
-        let source_paths: Vec<PathBuf> = raw_paths.iter().map(|s| {
-            let p = PathBuf::from(s);
-            if p.is_absolute() { p } else { workspace_root.join(s) }
-        }).collect();
-
-        let sem = Arc::clone(&self.parse_sem);
-        let mut new_library_uris: Vec<String> = Vec::new();
-        let mut all_results: Vec<FileIndexResult> = Vec::new();
-
-        for source_path in &source_paths {
-            if !source_path.exists() {
-                log::warn!("sourcePaths: {:?} does not exist, skipping", source_path);
-                continue;
-            }
-            log::info!("Indexing source path: {}", source_path.display());
-
-            let files = find_source_files_unconstrained(source_path);
-            log::info!("  Found {} source files in {}", files.len(), source_path.display());
-
-            let mut tasks = Vec::new();
-            for path in files {
-                let uri = match Url::from_file_path(&path) {
-                    Ok(u) => u,
-                    Err(_) => continue,
-                };
-                let uri_str = uri.to_string();
-                // Only tag as library if the file is OUTSIDE the workspace root.
-                // Files inside the workspace are already in the main index; sourcePaths
-                // can be used to un-ignore them without misclassifying them as libraries.
-                if !path.starts_with(&workspace_root) {
-                    new_library_uris.push(uri_str.clone());
-                }
-                let sem2 = Arc::clone(&sem);
-                let task: tokio::task::JoinHandle<Option<FileIndexResult>> = tokio::spawn(async move {
-                    let _permit = sem2.acquire_owned().await.ok()?;
-                    let content = tokio::fs::read_to_string(&path).await.ok()?;
-                    Some(Indexer::parse_file(&uri, &content))
-                });
-                tasks.push(task);
-            }
-
-            for task in tasks {
-                if let Ok(Some(result)) = task.await {
-                    all_results.push(result);
-                }
-            }
-        }
-
-        // Bail if workspace switched during async I/O.
-        if self.root_generation.load(Ordering::SeqCst) != gen {
-            log::info!("index_source_paths: generation changed during async I/O, discarding results");
-            return;
-        }
-
-        // Apply results additively (no reset_index_state).
-        for result in all_results {
-            let contrib = file_contributions(&result);
-            self.apply_contributions(contrib);
-        }
-
-        for uri in new_library_uris {
-            self.library_uris.insert(uri);
-        }
-
-        self.rebuild_bare_name_cache();
-        log::info!(
-            "Source paths indexed: {} library files, {} total indexed files",
-            self.library_uris.len(), self.files.len()
-        );
-    }
-
-    /// Primitive: drain a `FileContributions` into the DashMaps.
-    /// Deduplicates before inserting (same behaviour as before).
-    fn apply_contributions(&self, contrib: FileContributions) {
-        let (uri_str, file_data) = contrib.file_data;
-        let (hash_key, hash_val) = contrib.content_hash;
-
-        self.content_hashes.insert(hash_key, hash_val);
-        self.files.insert(uri_str.clone(), file_data);
-
-        for (name, locs) in contrib.definitions {
-            let mut entry = self.definitions.entry(name).or_default();
-            for loc in locs {
-                if !entry.iter().any(|l| l.uri == loc.uri && l.range == loc.range) {
-                    entry.push(loc);
-                }
-            }
-        }
-
-        for (key, loc) in contrib.qualified {
-            self.qualified.insert(key, loc);
-        }
-
-        for (pkg, uris) in contrib.packages {
-            let mut entry = self.packages.entry(pkg).or_default();
-            for u in uris {
-                if !entry.contains(&u) {
-                    entry.push(u);
-                }
-            }
-        }
-
-        for (super_name, locs) in contrib.subtypes {
-            let mut entry = self.subtypes.entry(super_name).or_default();
-            for loc in locs {
-                if !entry.iter().any(|l| l.uri == loc.uri && l.range == loc.range) {
-                    entry.push(loc);
-                }
-            }
-        }
-    }
-
-    /// Coordinator: rebuild bare-name cache from current definitions map.
-    pub fn rebuild_bare_name_cache(&self) {
-        if let Ok(mut cache) = self.bare_name_cache.write() {
-            *cache = build_bare_names(&self.definitions);
-        }
-        self.rebuild_importable_fqns();
-    }
-
-    /// Build importable_fqns: simple_name → [FQN, …] from real top-level symbols.
-    /// Uses files+package rather than the `qualified` map to avoid synthetic FileStem keys.
-    fn rebuild_importable_fqns(&self) {
-        let mut map: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
-        for file_entry in self.files.iter() {
-            let data = file_entry.value();
-            let pkg = match &data.package {
-                Some(p) if !p.is_empty() => p.clone(),
-                _ => continue,
-            };
-            // Detect top-level symbols: a symbol is top-level if its range is not
-            // wholly contained within any other symbol's range in the same file.
-            let syms = &data.symbols;
-            for (i, sym) in syms.iter().enumerate() {
-                let is_nested = syms.iter().enumerate().any(|(j, other)| {
-                    j != i
-                        && other.range.start.line <= sym.range.start.line
-                        && other.range.end.line >= sym.range.end.line
-                        && !(other.range.start.line == sym.range.start.line
-                            && other.range.end.line == sym.range.end.line)
-                });
-                if !is_nested {
-                    let fqn = format!("{}.{}", pkg, sym.name);
-                    map.entry(sym.name.clone()).or_default().push(fqn);
-                }
-            }
-        }
-        for fqns in map.values_mut() {
-            fqns.sort_unstable();
-            fqns.dedup();
-        }
-        if let Ok(mut guard) = self.importable_fqns.write() {
-            *guard = map;
-        }
-    }
-
-    /// (Re-)parse and index a single file's content in-place.
-    /// Returns immediately (no-op) when content is identical to the last indexed version.
-    /// Parse and index a single file.  Returns `Some(data)` when the file was
-    /// actually (re-)parsed, or `None` when the content-hash matched the previous
-    /// parse (no work done).  Callers that need to publish diagnostics should
-    /// read `data.syntax_errors` from the returned value.
-    pub fn index_content(&self, uri: &Url, content: &str) -> Option<Arc<FileData>> {
-        // Fast-path: skip re-parse if content hasn't changed since last index.
-        let hash = hash_str(content);
-        let uri_str = uri.to_string();
-        if self.content_hashes.get(&uri_str).map(|h| *h == hash).unwrap_or(false) {
-            return None;
-        }
-        
-        self.parse_count.fetch_add(1, Ordering::Relaxed);
-        // Invalidate cached completion items — the file is changing.
-        self.completion_cache.remove(&uri_str);
-        if let Ok(mut last) = self.last_completion.lock() {
-            *last = None;
-        }
-        
-        // Use pure parse function
-        let result = Self::parse_file(uri, content);
-        
-        // Apply result to index
-        self.apply_file_result(&result);
-        
-        // Rebuild bare-name cache so complete_bare doesn't iterate definitions.
-        self.rebuild_bare_name_cache();
-
-        Some(Arc::new(result.data))
-    }
-
-    /// Spawn background tasks to pre-warm the completion cache for all types
-    /// declared in `uri` as constructor parameters or properties.
-    ///
-    /// This runs after `index_content` so that when the user types `repo.` the
-    /// cache is already populated and the response is instant.
-    pub fn prewarm_completion_cache(self: Arc<Self>, uri: &Url) {
-        let Some(data) = self.files.get(uri.as_str()) else { return };
-        let from_uri = uri.clone();
-
-        // Collect unique type names from this file's lines.
-        let mut type_names: Vec<String> = Vec::new();
-        {
-            let mut seen = std::collections::HashSet::new();
-            for line in data.lines.iter() {
-                let t = line.trim_start();
-                if t.starts_with("//") || t.starts_with('*') { continue; }
-                let mut rest = t;
-                while let Some(ci) = rest.find(':') {
-                    let after = rest[ci + 1..].trim_start();
-                    let type_name: String = after.chars()
-                        .take_while(|&c| c.is_alphanumeric() || c == '_')
-                        .collect();
-                    if !type_name.is_empty()
-                        && type_name.chars().next().map(|c| c.is_uppercase()).unwrap_or(false)
-                        && seen.insert(type_name.clone())
-                    {
-                        type_names.push(type_name);
-                    }
-                    rest = &rest[ci + 1..];
-                }
-            }
-        }
-        drop(data);
-
-        // Spawn a background task per type (capped to avoid bursts).
-        let limit = Arc::new(tokio::sync::Semaphore::new(4));
-        for type_name in type_names {
-            // Skip if already cached or a primitive/stdlib type.
-            let idx = Arc::clone(&self);
-            let uri2 = from_uri.clone();
-            let sem = Arc::clone(&limit);
-            tokio::spawn(async move {
-                let _permit = sem.acquire_owned().await;
-                tokio::task::spawn_blocking(move || {
-                    // Resolve the type to a file URI.
-                    let locs = crate::resolver::resolve_symbol(&idx, &type_name, None, &uri2);
-                    if let Some(loc) = locs.first() {
-                        let file_uri = loc.uri.to_string();
-                        // Only bother if not already cached.
-                        if idx.completion_cache.contains_key(&file_uri) { return; }
-                        // This call builds + caches the items.
-                        crate::resolver::symbols_from_uri_as_completions_pub(&idx, &file_uri);
-                    }
-                }).await.ok();
-            });
-        }
-    }
-    ///
     /// LSP positions are UTF-16; for ASCII-heavy Kotlin/Java identifiers the
     /// character offset is identical to the UTF-16 unit offset.
     pub fn word_at(&self, uri: &Url, position: Position) -> Option<String> {
@@ -1466,18 +1018,6 @@ fn extract_class_decl_name(line: &str) -> Option<String> {
     Some(name)
 }
 
-// ─── hash helper ─────────────────────────────────────────────────────────────
-
-/// Fast non-cryptographic hash of a string (FNV-1a, no extra dependencies).
-fn hash_str(s: &str) -> u64 {
-    let mut h: u64 = 0xcbf2_9ce4_8422_2325;
-    for b in s.bytes() {
-        h ^= b as u64;
-        h = h.wrapping_mul(0x0000_0100_0000_01b3);
-    }
-    h
-}
-
 // ─── rg cross-file fallback ──────────────────────────────────────────────────
 
 // ─── misc helpers ────────────────────────────────────────────────────────────
@@ -1633,78 +1173,6 @@ mod tests {
             hover.contains("fun"),
             "Kotlin hover should say 'fun', got: {hover}"
         );
-    }
-
-    #[test]
-    fn parse_file_returns_symbols() {
-        let src = r#"
-package com.example
-
-class Foo {
-    fun bar(): String = "test"
-}
-"#;
-        let u = uri("/Foo.kt");
-        let result = Indexer::parse_file(&u, src);
-        
-        assert_eq!(result.uri, u);
-        assert_eq!(result.data.package, Some("com.example".to_string()));
-        assert_eq!(result.data.symbols.len(), 2); // class + fun
-        assert!(result.data.symbols.iter().any(|s| s.name == "Foo"));
-        assert!(result.data.symbols.iter().any(|s| s.name == "bar"));
-        assert!(result.error.is_none());
-    }
-
-    #[test]
-    fn parse_file_extracts_supertypes() {
-        let src = r#"
-interface Base
-class Child : Base
-"#;
-        let u = uri("/Child.kt");
-        let result = Indexer::parse_file(&u, src);
-        
-        // Debug: print what we found
-        eprintln!("Supertypes found: {:?}", result.supertypes);
-        
-        // Should find Child implements Base (interface Base has no supertypes)
-        assert!(result.supertypes.iter().any(|(name, _)| name == "Base"));
-    }
-
-    #[test]
-    fn apply_file_result_updates_index() {
-        let src = "class TestClass";
-        let u = uri("/Test.kt");
-        let result = Indexer::parse_file(&u, src);
-        
-        let idx = Indexer::new();
-        idx.apply_file_result(&result);
-        
-        // Verify symbol indexed
-        assert!(idx.definitions.contains_key("TestClass"));
-        let locs = idx.definitions.get("TestClass").unwrap();
-        assert_eq!(locs.len(), 1);
-        assert_eq!(locs[0].uri, u);
-    }
-
-    #[test]
-    fn apply_file_result_clears_stale_entries() {
-        let u = uri("/Test.kt");
-        
-        // First parse: class OldName
-        let result1 = Indexer::parse_file(&u, "class OldName");
-        let idx = Indexer::new();
-        idx.apply_file_result(&result1);
-        assert!(idx.definitions.contains_key("OldName"));
-        
-        // Second parse: class NewName (same file)
-        let result2 = Indexer::parse_file(&u, "class NewName");
-        idx.apply_file_result(&result2);
-        
-        // OldName should be gone
-        assert!(!idx.definitions.contains_key("OldName") || 
-                idx.definitions.get("OldName").unwrap().is_empty());
-        assert!(idx.definitions.contains_key("NewName"));
     }
 
     // ── word_at ──────────────────────────────────────────────────────────────
@@ -3545,114 +3013,7 @@ internal class ContactAddressInteractor @Inject constructor(
         assert_eq!(names, vec!["Apple", "Zebra"]);
     }
 
-    #[test]
-    fn apply_file_result_removes_both_stale_qualified_keys() {
-        let idx = Indexer::new();
-        let u = Url::parse("file:///pkg/Foo.kt").unwrap();
-        // First index: class Bar in file Foo.kt
-        idx.index_content(&u, "package com.example\nclass Bar {}");
-        assert!(idx.qualified.contains_key("com.example.Bar"), "initial pkg.Sym missing");
-        assert!(idx.qualified.contains_key("com.example.Foo.Bar"), "initial pkg.Stem.Sym missing");
-        // Re-index with Bar removed
-        idx.index_content(&u, "package com.example\n// empty");
-        assert!(!idx.qualified.contains_key("com.example.Bar"), "stale pkg.Sym not removed");
-        assert!(!idx.qualified.contains_key("com.example.Foo.Bar"), "stale pkg.Stem.Sym not removed");
-    }
-
     // ── apply_workspace_result ────────────────────────────────────────────────
-
-    /// apply_workspace_result must include files from cache hits (not just parsed files).
-    #[test]
-    fn apply_workspace_result_includes_cached_files() {
-        let u = uri("/Cached.kt");
-        // Build cache entry from a prior parse.
-        let parsed = Indexer::parse_file(&u, "class CachedClass");
-        let entry = FileCacheEntry {
-            mtime_secs: 0,
-            file_size: 0,
-            content_hash: 0,
-            file_data: parsed.data.clone(),
-        };
-        let cached = super::cache_entry_to_file_result(&u, &entry);
-
-        let workspace_result = WorkspaceIndexResult {
-            files: vec![cached],
-            stats: IndexStats { cache_hits: 1, ..Default::default() },
-            workspace_root: std::path::PathBuf::from("/"),
-            aborted: false,
-            complete_scan: true,
-        };
-
-        let idx = Indexer::new();
-        idx.apply_workspace_result(&workspace_result);
-
-        assert!(idx.definitions.contains_key("CachedClass"),
-            "CachedClass from cache hit should be in definitions index");
-        assert!(idx.files.contains_key(u.as_str()),
-            "CachedClass file should be in files map");
-    }
-
-    /// apply_workspace_result must do a full reset — switching workspaces removes
-    /// all symbols from the previous workspace.
-    #[test]
-    fn apply_workspace_result_clears_stale_workspace() {
-        let idx = Indexer::new();
-
-        let u1 = uri("/A.kt");
-        idx.apply_workspace_result(&WorkspaceIndexResult {
-            files: vec![Indexer::parse_file(&u1, "class ClassA")],
-            stats: IndexStats::default(),
-            workspace_root: std::path::PathBuf::from("/workspace_a"),
-            aborted: false,
-            complete_scan: true,
-        });
-        assert!(idx.definitions.contains_key("ClassA"), "ClassA should be present after first apply");
-
-        let u2 = uri("/B.kt");
-        idx.apply_workspace_result(&WorkspaceIndexResult {
-            files: vec![Indexer::parse_file(&u2, "class ClassB")],
-            stats: IndexStats::default(),
-            workspace_root: std::path::PathBuf::from("/workspace_b"),
-            aborted: false,
-            complete_scan: true,
-        });
-
-        assert!(!idx.definitions.contains_key("ClassA"),
-            "ClassA should be gone after workspace switch");
-        assert!(idx.definitions.contains_key("ClassB"),
-            "ClassB should be present in new workspace");
-    }
-
-    /// apply_workspace_result must combine both cache hits and freshly parsed files
-    /// into the final index — neither should be silently dropped.
-    #[test]
-    fn apply_workspace_result_mixed_cache_and_parsed() {
-        let u_cached = uri("/Cached.kt");
-        let u_parsed  = uri("/Parsed.kt");
-
-        let cached_parse = Indexer::parse_file(&u_cached, "class CachedClass");
-        let entry = FileCacheEntry {
-            mtime_secs: 0, file_size: 0, content_hash: 0, file_data: cached_parse.data.clone()
-        };
-        let cached_result = super::cache_entry_to_file_result(&u_cached, &entry);
-
-        let parsed_result = Indexer::parse_file(&u_parsed, "class ParsedClass");
-
-        let idx = Indexer::new();
-        idx.apply_workspace_result(&WorkspaceIndexResult {
-            files: vec![cached_result, parsed_result],
-            stats: IndexStats { cache_hits: 1, files_parsed: 1, ..Default::default() },
-            workspace_root: std::path::PathBuf::from("/"),
-            aborted: false,
-            complete_scan: true,
-        });
-
-        assert!(idx.definitions.contains_key("CachedClass"),
-            "CachedClass (from cache) should be in index");
-        assert!(idx.definitions.contains_key("ParsedClass"),
-            "ParsedClass (freshly parsed) should be in index");
-        assert_eq!(idx.files.len(), 2, "exactly 2 files in index");
-    }
 
     #[test]
     fn debug_super_chain() {
